requirements.txt
Django==6.0.3
python-decouple
dj-database-url
django-cloudinary-storage
cloudinary
psycopg2-binary
